refactor: remove helmut.connector.base
